agent/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.52.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
pyyaml==6.0.2
requests==2.32.3
faiss-cpu==1.9.0
rank_bm25==0.2.2
numpy==1.26.4
python-multipart==0.0.12
